requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coverage==7.4.3
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.9
exceptiongroup==1.2.0
filelock==3.9.0
flake8==7.0.0
Flask==3.0.2
fonttools==4.49.0
fsspec==2023.4.0
idna==3.6
importlib_metadata==7.0.2
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
kiwisolver==1.4.5
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.3
-e git+https://github.com/open-mmlab/mmpose.git@5a3be9451bdfdad2053a90dc1199e3ff1ea1a409#egg=mmpose
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.2.1
parameterized==0.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
py==1.11.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.0.2
pytest-runner==6.0.1
python-dateutil==2.9.0
pytorchcv==0.0.67
pytz==2023.4
PyYAML==6.0.1
requests==2.28.2
rich==13.4.2
rmn==3.1.2
scikit-learn==1.2.2
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.3.0
toml==0.10.2
tomli==2.0.1
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.65.2
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.1
xdoctest==1.1.3
xtcocotools==1.14.3
yapf==0.40.2
zipp==3.17.0
